% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{compute_edge_weights}
\alias{compute_edge_weights}
\title{compute_edge_weights}
\usage{
compute_edge_weights(CTC, baseline)
}
\arguments{
\item{CTC}{.}

\item{baseline}{.}
}
\value{
Edge weights.
}
\description{
compute_edge_weights
}
\keyword{internal}
