% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{compute_edge_weights_for_p_values}
\alias{compute_edge_weights_for_p_values}
\title{compute_edge_weights_for_p_values}
\usage{
compute_edge_weights_for_p_values(p_value, p_value_cutoff)
}
\arguments{
\item{p_value}{.}

\item{p_value_cutoff}{.}
}
\value{
Edge weights.
}
\description{
compute_edge_weights_for_p_values
}
\keyword{internal}
