% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{end_doFuture_strategy}
\alias{end_doFuture_strategy}
\title{End the currently set doFuture strategy}
\usage{
end_doFuture_strategy()
}
\description{
Resets the default strategy using \code{future::plan("default")}.
}
\keyword{internal}
