% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{Extreme_causality_graph}
\alias{Extreme_causality_graph}
\title{Extreme Granger-causal graph estimate}
\usage{
Extreme_causality_graph(
  w,
  max_causal_lag = 1,
  max_confounder_lag = 0,
  nu_x = 0.3,
  q_y = 0.2,
  q_z = 0.1,
  instant = FALSE,
  both_tails = TRUE,
  p_value_based = FALSE,
  p_value_cutoff = 0.05
)
}
\arguments{
\item{w}{A \code{data.frame} of all time series, which should be numeric and of the same length.}

\item{max_causal_lag}{The time delay for the effect from \code{x} to \code{y}. This is the coefficient 'p' in Appendix A of the manuscript.}

\item{max_confounder_lag}{The lag from \eqn{Z} to \eqn{(X, Y)}. If the common cause has different lags to \eqn{X} and \eqn{Y}, it may cause spurious causality between \eqn{X} and \eqn{Y}.
Ensure \code{max_confounder_lag} is larger than this lag.}

\item{nu_x}{The coefficient \eqn{\tau_X} or \eqn{k_n} in the manuscript, defined as \eqn{k_n = \lfloor n^{\nu_x} \rfloor}. If strong hidden confounding is expected, set \code{nu_x} to 0.4 or 0.5.}

\item{q_y}{The coefficient \eqn{\tau_y = q_y \times n}, describing the conditioning on \eqn{Y_t}. For large auto-correlation in \eqn{Y}, set \code{q_y} to 0.1 or less. Note that in the manuscript, \eqn{q_y} is defined as \code{1 - q_y}.}

\item{q_z}{The coefficient \eqn{\tau_z = q_z \times n}, describing the conditioning on \eqn{Z_t}. This is irrelevant if \code{z} is \code{NULL}.
For strong confounding effects, set \code{q_z} to 0.2 or 0.3. Note that in the manuscript, \eqn{q_z} is defined as \code{1 - q_z}.}

\item{instant}{Whether instantaneous effects should be captured; defaults to \code{FALSE}.}

\item{both_tails}{Set to \code{TRUE} to consider both large and extremely negative values.
For example, in GARCH models, both tails are of interest, while in VAR models, only large values might be relevant.}

\item{p_value_based}{If \code{FALSE}, Algorithm 1 is used for inferring the edges.
If \code{TRUE}, the testing procedure with a cut-off p-value of \code{p_value_cutoff} is used for detecting the presence of an edge.
These procedures typically output similar results, but the testing procedure is significantly slower.}

\item{p_value_cutoff}{P-value cut-off level to reject the absence of an edge in the estimated graph.}
}
\value{
A named list containing:
\item{G}{A graph defined by its edges. Each row corresponds to an edge from the first column pointing to the second column.
Use \code{graph <- graph_from_edgelist(G$G)} from the \link[igraph:igraph]{igraph} package to obtain the graph environment;}
\item{weights}{Weights corresponding to each edge, representing how close the coefficient \eqn{\hat{\Gamma}_{X\rightarrow Y | Z}} is to 1.
If \eqn{\hat{\Gamma}_{X\rightarrow Y | Z} = 1}, the weight is 1.
The weight is 0 if \eqn{\hat{\Gamma}_{X\rightarrow Y | Z} = (1 + \hat{\Gamma}^{baseline}_{X\rightarrow Y | Z}) / 2}.}
}
\description{
This function estimates the causal graph (path diagram) between a set of time series.
Use \code{\link[=Extreme_causality_graph_parallel]{Extreme_causality_graph_parallel()}} instead, for faster parallel computations.
}
