% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{Extreme_causality_test}
\alias{Extreme_causality_test}
\title{Extreme Granger-causal test}
\usage{
Extreme_causality_test(
  x,
  y,
  z = NULL,
  max_causal_lag = 1,
  max_confounder_lag = 0,
  nu_x = 0.3,
  q_y = 0.2,
  q_z = 0.1,
  both_tails = TRUE,
  instant = FALSE,
  p_value_computation = FALSE,
  bootstrap_repetitions = 50,
  choice_of_F = 0.5
)
}
\arguments{
\item{x}{A numeric vector representing the first time series (potential cause).}

\item{y}{A numeric vector representing the second time series (potential effect).}

\item{z}{A \code{data.frame} of potential confounders. Set to \code{NULL} if there are no confounders.}

\item{max_causal_lag}{The time delay for the effect from \code{x} to \code{y}. This is the coefficient 'p' in Appendix A of the manuscript.}

\item{max_confounder_lag}{The lag from \eqn{Z} to \eqn{(X, Y)}. If the common cause has different lags to \eqn{X} and \eqn{Y}, it may cause spurious causality between \eqn{X} and \eqn{Y}.
Ensure \code{max_confounder_lag} is larger than this lag.}

\item{nu_x}{The coefficient \eqn{\tau_X} or \eqn{k_n} in the manuscript, defined as \eqn{k_n = \lfloor n^{\nu_x} \rfloor}. If strong hidden confounding is expected, set \code{nu_x} to 0.4 or 0.5.}

\item{q_y}{The coefficient \eqn{\tau_y = q_y \times n}, describing the conditioning on \eqn{Y_t}. For large auto-correlation in \eqn{Y}, set \code{q_y} to 0.1 or less. Note that in the manuscript, \eqn{q_y} is defined as \code{1 - q_y}.}

\item{q_z}{The coefficient \eqn{\tau_z = q_z \times n}, describing the conditioning on \eqn{Z_t}. This is irrelevant if \code{z} is \code{NULL}.
For strong confounding effects, set \code{q_z} to 0.2 or 0.3. Note that in the manuscript, \eqn{q_z} is defined as \code{1 - q_z}.}

\item{both_tails}{Set to \code{TRUE} to consider both large and extremely negative values.
For example, in GARCH models, both tails are of interest, while in VAR models, only large values might be relevant.}

\item{instant}{Whether instantaneous effects should be captured; defaults to \code{FALSE}.}

\item{p_value_computation}{If set to \code{FALSE} the faster "Algorithm 1" is used.
If set to \code{TRUE} the p-value for the hypothesis \eqn{H_0: X \text{ does not cause } Y \text{ in extremes given } Z} is computed.
If \code{p_value < 0.05}, we conclude that \eqn{X} causes \eqn{Y} given \eqn{Z}.}

\item{bootstrap_repetitions}{The number of bootstrap repetitions for p-value computation. More repetitions yield more precise p-values but require longer computation time.}

\item{choice_of_F}{Choice of F in the coefficient. Leave default unless you want to reproduce the results from the manuscript}
}
\value{
A named list containing:
\item{is_causal}{A logical value indicating whether evidence of causality is detected;}
\item{output}{Either 'Evidence of causality' or 'No causality' based on Algorithm 1 from the manuscript;}
\item{p_value_tail}{This is not shown if \code{p_value_computation==FALSE}. Rejection indicates evidence of causality in tail.
It corresponds to the p-value for the hypothesis H_0: "X does not cause Y in tail given Z", based on bootstrapping. Often \code{p_value==1} which means that \code{CTC<baseline};}
\item{p_value_extreme}{This is not shown if \code{p_value_computation==FALSE}. Rejection indicates evidence of causality in extremes.
It corresponds to the p-value for the hypothesis H_0: \eqn{\hat{\Gamma}_{X\rightarrow Y | Z} < (1 + 3 \cdot \hat{\Gamma}^{baseline}_{X\rightarrow Y | Z}) / 4};}
\item{CTC}{The coefficient \eqn{\hat{\Gamma}_{X\rightarrow Y | Z}};}
\item{baseline}{The baseline coefficient \eqn{\hat{\Gamma}^{baseline}_{X\rightarrow Y | Z}}.}
}
\description{
This function tests whether the tails/extremes of a time series \code{X} cause those of a time series \code{Y}, given potential confounders \code{Z}.
}
