% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtremeGranger-package.R
\docType{package}
\name{ExtremeGranger-package}
\alias{ExtremeGranger}
\alias{ExtremeGranger-package}
\title{ExtremeGranger: Granger Causality in Extremes}
\description{
This framework for Granger causality in extremes is designed to identify causal links from extreme events in observational time series. Granger causality plays a pivotal role in understanding directional relationships among time-varying variables. While the task of causal discovery in time series gains heightened importance during extreme and highly volatile periods, state-of-the-art methods primarily focus on causality within the body of the distribution, often overlooking causal mechanisms that manifest only during extreme events. Our framework is designed to infer causality mainly from extreme events by leveraging the causal tail coefficient. It is especially helpful in the presence of hidden confounders. Our method is non-parametric, can handle non-linear and high-dimensional time series. The methodology was introduced in Bodik and Pasche (2024) \doi{10.48550/arXiv.2407.09632}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://opasche.github.io/ExtremeGranger/}
  \item \url{https://github.com/opasche/ExtremeGranger}
  \item Report bugs at \url{https://github.com/opasche/ExtremeGranger/issues}
}

}
\author{
\strong{Maintainer}: Olivier C. Pasche \email{olivier_pasche@alumni.epfl.ch} (\href{https://orcid.org/0000-0002-1202-9199}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Olivier C. Pasche \email{olivier_pasche@alumni.epfl.ch} (\href{https://orcid.org/0000-0002-1202-9199}{ORCID}) [copyright holder]
  \item Juraj Bodik \email{juraj.bodik@unil.ch} (\href{https://orcid.org/0000-0002-4970-3300}{ORCID}) [copyright holder]
}

}
\keyword{internal}
