% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{find_parents}
\alias{find_parents}
\title{find_parents}
\usage{
find_parents(G, vertex)
}
\arguments{
\item{G}{.}

\item{vertex}{.}
}
\value{
Parents
}
\description{
find_parents
}
\keyword{internal}
