% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{run_future_max}
\alias{run_future_max}
\title{run_future_max}
\usage{
run_future_max(x, k = 3, instant = TRUE)
}
\arguments{
\item{x}{.}

\item{k}{.}

\item{instant}{.}
}
\value{
.
}
\description{
run_future_max
}
\keyword{internal}
