% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{switcher_for_test}
\alias{switcher_for_test}
\title{switcher_for_test}
\usage{
switcher_for_test(x, number_of_blocks = 15)
}
\arguments{
\item{x}{.}

\item{number_of_blocks}{.}
}
\value{
.
}
\description{
switcher_for_test
}
\keyword{internal}
